Move code from config/chains to config/tvl
